Fix memory-map inside qemu for mimasv2
Qemu wasn't passing a pre-crt test that was being executed in the jcore bootrom in _testbra. A value of -117 was being written to a memory location outside fo the memory map. The write was being silently ignored. A subsequent read-back of the unmapped  memory location returned 0, which caused the test to fail.

The map now accurately reflects the memory layout.

SRAM: 32kB @ 0x00000000
DRAM: 64MB @ 0x01000000

The DRAM is unused by the bootrom. However, a different bootrom revision that did use the DRAM and left SRAM read-only succeeded in passing the test as well. Thus the DRAM mapping is functional.