Do not unnecessarily clear Q M and T bits in cpu_sr
Previously, in the _testmov subroutine of the J2 test suite, I would write a
value of 0x3f3 to the Status Register, and then immediately read it back, but
the Q bit was (for some reason) not set. This was because the gen_write_sr()
function was clearing the Q M and T bits in the cpu_sr variable.
The Q M and T bits are broken out of cpu_sr presumably for efficency purposes,
into cpu_sr_q, cpu_sr_m, and cpu_sr_t. Since gen_read_sr() ORs the individual,
broken-out bits with the value in cpu_sr, it was unnecessary to mask them out.
However, masking them out and re-orring them with the individual bits was
producing inconsistent results, causing the _testmov subroutine to fail. This
at least allows me to get past the _testmov subroutine. I believe that this
change is also 'safe' otherwise.

I'm not sure how smart it was on the part of the original SH4 port author
to have several copies of SR bits floating around. It makes maintainence more
difficult, but could potentially improve performance.