Working for cortex-m3 target
It would still be helpful to test on e.g. cortex-m0, cortex-m0+, cortex-m4, cortex-m33, etc. Not sure if there are any cortex-r targets that are emulated at the moment.

The cortex-m0 / cortex-m0+ targets do not support the primask or faultmask registers.

The exact same procedure was used in 2015 to test the code, but the patchset is far smaller this time around.

https://www.mail-archive.com/qemu-devel@nongnu.org/msg340867.html