Added stm32f4xx-fmc memory mapped device (unimplemented)
This change adds basic support for the memory-mapped registers of the flexible memory controller (fmc) on the stm32f4xx series of microcontrollers.

Currently there is no effect of writing to the controller address space. The only purpose of having this unimplemented controller in the address space is to prevent qemu from crashing if the
guest OS writes to the address space (e.g. if the guest OS would like to configure the SDRAM controller).